fix validate admin, fix homepage, login register, them 1 so cac chuc nang co ban, sample

+++ src/main/java/com/phonecard/model/CardProduct.java
@@ -1,3 +1,8 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+
 package com.phonecard.model;
 
 import java.math.BigDecimal;
@@ -6,37 +11,28 @@ public class CardProduct {
     private int productId;
     private int providerId;
     private String productName;
-    private BigDecimal faceValue;    
-    private BigDecimal sellingPrice; 
+    private double faceValue;
+    private double sellingPrice;
     private String description;
-    
-    private Provider provider; 
-    
+    private Provider provider;
     private int availableCount; 
 
     public CardProduct() {}
 
     public int getProductId() { return productId; }
     public void setProductId(int productId) { this.productId = productId; }
-
     public int getProviderId() { return providerId; }
     public void setProviderId(int providerId) { this.providerId = providerId; }
-
     public String getProductName() { return productName; }
     public void setProductName(String productName) { this.productName = productName; }
-
-    public BigDecimal getFaceValue() { return faceValue; }
-    public void setFaceValue(BigDecimal faceValue) { this.faceValue = faceValue; }
-
-    public BigDecimal getSellingPrice() { return sellingPrice; }
-    public void setSellingPrice(BigDecimal sellingPrice) { this.sellingPrice = sellingPrice; }
-
+    public double getFaceValue() { return faceValue; }
+    public void setFaceValue(double faceValue) { this.faceValue = faceValue; }
+    public double getSellingPrice() { return sellingPrice; }
+    public void setSellingPrice(double sellingPrice) { this.sellingPrice = sellingPrice; }
     public String getDescription() { return description; }
     public void setDescription(String description) { this.description = description; }
-
     public Provider getProvider() { return provider; }
     public void setProvider(Provider provider) { this.provider = provider; }
-
     public int getAvailableCount() { return availableCount; }
     public void setAvailableCount(int availableCount) { this.availableCount = availableCount; }
 }
